requirements.txt
Django>=4.2.0
python-dotenv>=1.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
Pillow>=10.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.2.0
psycopg2-binary>=2.9.7
gunicorn>=21.2.0
whitenoise>=6.5.0
pydicom
numpy>=1.20.0
pydicom>=2.3.0
SimpleITK>=2.1.0
rt-utils>=1.2.0
pynrrd>=0.4.2
requests
numpy>=1.20.0
pydicom>=2.3.0
SimpleITK>=2.1.0
rt-utils>=1.2.0
pynrrd>=0.4.2
git+https://github.com/AIM-Harvard/pyradiomics.git
cython
pandas
bokeh>=3.0.0
scipy>=1.9.0
numpy>=1.21.0
scikit-learn>=1.0.0
python-dotenv
reportlab